examples/microsoft_flask/requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
decorator==5.1.0
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msal==1.14.0
pycparser==2.20
PyJWT==2.1.0
pymojang==1.3.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
validators==0.18.2
Werkzeug==2.0.1
